# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [users]
      summary: Get the territory leaderboard
      description: Returns at most 20 runners ordered by totalHexesOwned descending, then total distance, run count, and user ID.
      parameters:
        - name: currentUserId
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UserId"
        - name: scope
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/LeaderboardScope"
      responses:
        "200":
          description: Territory leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResult"
        "500":
          description: Leaderboard could not be loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /users/{userId}/baseline:
    patch:
      operationId: updateUserBaseline
      tags: [users]
      summary: Save a runner's initial fitness baseline
      security:
        - anonymousDeviceCredential: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserBaselineUpdate"
      responses:
        "200":
          description: Baseline saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserBaseline"
        "400":
          description: Invalid baseline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Missing or invalid anonymous device credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Device credential does not match the runner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /users/{userId}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user totals and recent runs
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UserId"
      responses:
        "200":
          description: User totals and up to five recent runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStatsResult"
        "400":
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Stats could not be loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    anonymousDeviceCredential:
      type: http
      scheme: bearer
      bearerFormat: HexRunner anonymous device credential
  schemas:
...
    UserId:
      type: string
      minLength: 8
      maxLength: 120
      pattern: "^[A-Za-z0-9_-]+$"
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: number
          multipleOf: 1
          minimum: 1
        displayName:
          type: string
        totalHexesOwned:
          type: number
          multipleOf: 1
          minimum: 0
        totalRuns:
          type: number
          multipleOf: 1
          minimum: 0
        totalDistanceKm:
          type: number
          minimum: 0
        isCurrentUser:
          type: boolean
        totalCredits: { type: number, multipleOf: 1, minimum: 0 }
        totalBonusCredits: { type: number, multipleOf: 1, minimum: 0 }
      required: [rank, displayName, totalHexesOwned, totalRuns, totalDistanceKm, isCurrentUser, totalCredits, totalBonusCredits]
    LeaderboardResult:
      type: object
      properties:
        scope:
          $ref: "#/components/schemas/LeaderboardScope"
        users:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
      required: [scope, users]
    RecentRun:
      type: object
      properties:
        runId:
          type: string
        distanceKm:
          type: number
        averagePaceMinPerKm:
          type: [number, "null"]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        claimedHexes:
          type: number
          multipleOf: 1
          minimum: 0
        newHexes:
          type: number
          multipleOf: 1
          minimum: 0
        stolenHexes:
          type: number
          multipleOf: 1
          minimum: 0
        bonusCredit: { type: number, multipleOf: 1, minimum: 0 }
        totalCredit: { type: number, multipleOf: 1, minimum: 0 }
      required: [runId, distanceKm, averagePaceMinPerKm, startedAt, endedAt, claimedHexes, newHexes, stolenHexes, bonusCredit, totalCredit]
    UserStatsTotals:
      type: object
      properties:
        totalRuns:
          type: number
          multipleOf: 1
          minimum: 0
        totalDistanceKm:
          type: number
          minimum: 0
        totalElapsedSeconds:
          type: number
          multipleOf: 1
          minimum: 0
        averagePaceMinPerKm:
          type: [number, "null"]
        totalHexesOwned:
          type: number
          multipleOf: 1
          minimum: 0
        totalClaimedHexes:
          type: number
          multipleOf: 1
          minimum: 0
        totalNewHexes:
          type: number
          multipleOf: 1
          minimum: 0
        totalStolenHexes:
          type: number
          multipleOf: 1
          minimum: 0
        currentStreak:
          type: number
          multipleOf: 1
          minimum: 0
        todayClaimedHexes:
          type: number
          multipleOf: 1
          minimum: 0
        dailyBudget:
          type: number
          multipleOf: 1
          minimum: 1
        totalCredits: { type: number, multipleOf: 1, minimum: 0 }
        totalBonusCredits: { type: number, multipleOf: 1, minimum: 0 }
        todayBonusCredits: { type: number, multipleOf: 1, minimum: 0 }
        dailyBonusCap: { type: number, multipleOf: 1, minimum: 1 }
      required: [totalRuns, totalDistanceKm, totalElapsedSeconds, averagePaceMinPerKm, totalHexesOwned, totalClaimedHexes, totalNewHexes, totalStolenHexes, currentStreak, todayClaimedHexes, dailyBudget, totalCredits, totalBonusCredits, todayBonusCredits, dailyBonusCap]
...
    FitnessTier:
      type: string
      enum: [beginner, casual, regular, trained]
    LeaderboardScope:
      type: string
      enum: [global, city, friends]
    UserBaselineUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 40
        city:
          type: string
          minLength: 1
          maxLength: 60
        activityLevel:
          $ref: "#/components/schemas/FitnessTier"
      required: [city, activityLevel]
    UserBaseline:
      type: object
      properties:
        displayName:
          type: string
        city:
          type: string
        activityLevel:
          $ref: "#/components/schemas/FitnessTier"
        completedAt:
          type: string
          format: date-time
      required: [displayName, city, activityLevel, completedAt]
    TakeoverAlert:
      type: object
      properties:
        h3Index:
          type: string
        happenedAt:
          type: string
          format: date-time
      required: [h3Index, happenedAt]
    UserStatsResult:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        baseline:
          oneOf:
            - $ref: "#/components/schemas/UserBaseline"
            - type: "null"
        totals:
          $ref: "#/components/schemas/UserStatsTotals"
        recentRuns:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/RecentRun"
        takeoverAlerts:
          type: array
          maxItems: 3
          items:
            $ref: "#/components/schemas/TakeoverAlert"
      required: [userId, displayName, baseline, totals, recentRuns, takeoverAlerts]
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...