# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  /** @minimum 1 */
  rank: number;
  displayName: string;
  /** @minimum 0 */
  totalHexesOwned: number;
  /** @minimum 0 */
  totalRuns: number;
  /** @minimum 0 */
  totalDistanceKm: number;
  isCurrentUser: boolean;
  /** @minimum 0 */
  totalCredits: number;
  /** @minimum 0 */
  totalBonusCredits: number;
}