# lib/api-zod/src/generated/types/recentRun.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentRun {
  runId: string;
  distanceKm: number;
  /** @nullable */
  averagePaceMinPerKm: number | null;
  startedAt: Date;
  endedAt: Date;
  /** @minimum 0 */
  claimedHexes: number;
  /** @minimum 0 */
  newHexes: number;
  /** @minimum 0 */
  stolenHexes: number;
  /** @minimum 0 */
  bonusCredit: number;
  /** @minimum 0 */
  totalCredit: number;
}