# lib/api-zod/src/generated/types/userStatsTotals.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserStatsTotals {
  /** @minimum 0 */
  totalRuns: number;
  /** @minimum 0 */
  totalDistanceKm: number;
  /** @minimum 0 */
  totalElapsedSeconds: number;
  /** @nullable */
  averagePaceMinPerKm: number | null;
  /** @minimum 0 */
  totalHexesOwned: number;
  /** @minimum 0 */
  totalClaimedHexes: number;
  /** @minimum 0 */
  totalNewHexes: number;
  /** @minimum 0 */
  totalStolenHexes: number;
  /** @minimum 0 */
  currentStreak: number;
  /** @minimum 0 */
  todayClaimedHexes: number;
  /** @minimum 1 */
  dailyBudget: number;
  /** @minimum 0 */
  totalCredits: number;
  /** @minimum 0 */
  totalBonusCredits: number;
  /** @minimum 0 */
  todayBonusCredits: number;
  /** @minimum 1 */
  dailyBonusCap: number;
}